[Part 2] Reorder code in Icd Managenet Cluster.
